Generation of certificates and keys for etcd gated if etcd is disabled.
Problem:
When support for etcd was added in 3957142, generation of certificates
and keys for etcd was not gated behind use of managed etcd.
Keys are generated and distributed across servers even if managed etcd
is not enabled.

Solution:
Allow generation of certificates and keys only if managed etc is
enabled. Check config.DisableETCD flag.

Signed-off-by: Bartossh <lenartconsulting@gmail.com>
Signed-off-by: Derek Nola <derek.nola@suse.com>
Co-authored-by: Bartosz Lenart <lenart.consulting@gmail.com>
porting by
Signed-off-by: Deshi Xiao <xiaods@gmail.com>